feat(lifecycle): distinguish between hard and soft exits
## Overview
The Lifecycle API needs a way for applications to do both soft exits and hard exits.

A soft exit moves the app to the background state, and returns the user to the home screen.

A hard exit moves the app through the closing, and closed states.

## Justification
This is to enable two use cases:

- As an app developer, I want my application to go in the `background` state when the user hits _home_ or _back_ on their remote, in case the user decices to come back quickly, or left the app accidentally.

- As an app developer, I want my application to go into the `closing` and then `closed` states when the user manually selects an "Exit" option from the menu, or for some operational reason like an update or bad state.

## Changes
- Propose we rename `close()` to `exit()`
- Add an optional `unload` (boolean) parameter to exit