DeepMerge failed to merge instantiated objects
I have observed that the DeepMerge library fails to correctly merge
instantiated objects, which is causing issues with the Stage object
context option. The canvas option value is being reassinged after
DeepMerge but the context option is not being handled
properly.

To resolve this issue, I have introduced a new option called
'isMergeableObject' in the DeepMerge library. This option copies the
whole instantiated object, instead of just copying its properties. This
allows the proper merging of instantiated objects and resolves both the
canvas and context options issues with DeepMerge.

Signed-off-by: suresh-gangumalla <sureshkumar.gangumalla@soctronics.com>